pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airlock-llm"
version = "0.1.0"
description = "LLM security gateway — PII redaction, prompt injection defense, rate limiting, and cost tracking as a drop-in reverse proxy."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "AI Engineering Toolkit Contributors"}]
keywords = ["llm", "security", "proxy", "pii", "prompt-injection", "gateway"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "httpx>=0.25",
    "structlog>=23.1",
    "tiktoken>=0.5",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "httpx",
    "ruff>=0.1",
    "mypy>=1.7",
    "types-PyYAML",
]

[project.scripts]
airlock = "airlock.cli:main"

[project.urls]
Homepage = "https://github.com/YOUR_ORG/airlock"
Documentation = "https://github.com/YOUR_ORG/airlock/tree/main/docs"
Repository = "https://github.com/YOUR_ORG/airlock"
Issues = "https://github.com/YOUR_ORG/airlock/issues"
Changelog = "https://github.com/YOUR_ORG/airlock/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
